requirements.txt
discord.py
SQLAlchemy
psycopg2